requirements.txt
streamlit
sentence-transformers
faiss-cpu
pandas
numpy
tqdm